+++ History.md
@@ -48,6 +48,7 @@
   * Prefer the rackup file specified by the CLI (#2225)
   * Fix for spawning subprocesses with fork_worker option (#2267)
   * Set `CONTENT_LENGTH` for chunked requests (#2287)
+  * Better error handling during force shutdown (#2271)
 
 * Refactor
   * Remove unused loader argument from Plugin initializer (#2095)